requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
motor==3.3.2
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
